feat: extract function
